Add best readme ever-est
This project was inspired by https://www.youtube.com/watch? v=Re3rijPsGoY video created by an iconic developer known all over the world. The decision for duplicating the idea was that the high life PC Master Race needs solution for the one of the world's most critical issues - is SOMETHING even or odd but the used technology can't be so peasant looking as JavaScript. That's why our team decided to choose a modern, refined, high performance and no boilerplate code technology. The only possible solution was god's favorite language - Java.  Ladies, gentleman and others we are proud to present you - Is it even or odd as a microservice!